Extract opening period details to its own component
